# api/openapi.yaml
openapi: 3.0.3
info:
  title: Drinkin' API
  description: Backend API for the Drinkin' social platform (drink experiences & reviews)
  version: "0.1.0"
servers:
  - url: https://api.drinkin.app/v1
    description: Production
  - url: http://localhost:8080/api
    description: Local development

tags:
  - name: Auth
  - name: Users
  - name: Posts
  - name: Feed
  - name: Social
  - name: Engagement

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error (e.g. underage, weak password)
        '409':
          description: Email or username already exists

  /auth/login:
    post:
      tags: [Auth]
      summary: Log in an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  /users/me:
    get:
      tags: [Users]
      summary: Get current authenticated user's profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    put:
      tags: [Users]
      summary: Update current user's profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

  /users/{id}:
    get:
      tags: [Users]
      summary: Get a user's public profile
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Public profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: User not found

  /users/{id}/follow:
    post:
      tags: [Social]
      summary: Follow a user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Followed successfully
        '400':
          description: Cannot follow yourself
    delete:
      tags: [Social]
      summary: Unfollow a user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Unfollowed successfully

  /posts:
    post:
      tags: [Posts]
      summary: Create a new post
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'

  /posts/{id}:
    get:
      tags: [Posts]
      summary: Get a single post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found

  /posts/{id}/reactions:
    post:
      tags: [Engagement]
      summary: React to a post
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactRequest'
      responses:
        '204':
          description: Reacted successfully (idempotent)
    delete:
      tags: [Engagement]
      summary: Remove own reaction from a post
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204':
          description: Reaction removed
    get:
      tags: [Engagement]
      summary: List reactions on a post
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Paginated reactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionPage'

  /posts/{id}/comments:
    get:
      tags: [Engagement]
      summary: List comments on a post
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Paginated comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPage'
    post:
      tags: [Engagement]
      summary: Add a comment to a post
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'

  /feed:
    get:
      tags: [Feed]
      summary: Get the current user's feed (posts from followed users)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Paginated feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPage'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PostId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor

  schemas:
    RegisterRequest:
      type: object
      required: [email, username, password, dateOfBirth]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 8
        dateOfBirth:
          type: string
          format: date

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserProfile'

    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        drinkPreferences:
          type: array
          items:
            type: string
        followerCount:
          type: integer
        followingCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        drinkPreferences:
          type: array
          items:
            type: string

    CreatePostRequest:
      type: object
      required: [text, drinkCategory]
      properties:
        text:
          type: string
          maxLength: 3000
        drinkCategory:
          type: string
          enum: [ALCOHOLIC, NON_ALCOHOLIC]
        drinkType:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        tastingNotes:
          type: string
        scenario:
          type: string
        mediaUrls:
          type: array
          items:
            type: string

    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author:
          $ref: '#/components/schemas/UserProfile'
        text:
          type: string
        drinkCategory:
          type: string
        drinkType:
          type: string
        rating:
          type: integer
        tastingNotes:
          type: string
        scenario:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    PostPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        nextCursor:
          type: string
          nullable: true

    ReactRequest:
      type: object
      required: [reactionType]
      properties:
        reactionType:
          type: string
          enum: [LIKE, LOVE, CHEERS, WOW, SAD]

    Reaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserProfile'
        reactionType:
          type: string
        createdAt:
          type: string
          format: date-time

    ReactionPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Reaction'
        nextCursor:
          type: string
          nullable: true

    CreateCommentRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          maxLength: 1000

    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
        author:
          $ref: '#/components/schemas/UserProfile'
        text:
          type: string
        createdAt:
          type: string
          format: date-time

    CommentPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        nextCursor:
          type: string
          nullable: true